# Optimization of Base58 Encode/Deocde: 2X-14X ++