* Port of https://github.com/egonelbre/btcutil/commits/perf

Before:
    BenchmarkBase58Decode_5K-32     266      4373774 ns/op  1.56 MB/s
    BenchmarkBase58Decode_100K-32   1     1516196700 ns/op  0.09 MB/s
    BenchmarkBase58Encode_5K-32     46      23934763 ns/op  0.21 MB/s
    BenchmarkBase58Encode_100K-32   1     9351948600 ns/op  0.01 MB/s
After:
    BenchmarkBase58Decode_5K-32     3868      277944 ns/op  24.57 MB/s
    BenchmarkBase58Decode_100K-32   13      83772100 ns/op   1.63 MB/s
    BenchmarkBase58Encode_5K-32     501       2419129 ns/op  2.07 MB/s
    BenchmarkBase58Encode_100K-32   2       923507950 ns/op  0.11 MB/s

Same code in use here, with similar improvement noted.