@johnsonjh 2020-11-16
This failure actually appears to be a failure with the Travis CI and unrelated to the patch.