@johnsonjh 2020-11-16
This is a false positive caused by Travis and is fixed in https://github.com/pkt-cash/pktd/pull/198